order api completed

+++ src/main/java/com/basu/trading/service/OrderServiceImpl.java
@@ -2,10 +2,7 @@ package com.basu.trading.service;
 
 import com.basu.trading.domain.OrderStatus;
 import com.basu.trading.domain.OrderType;
-import com.basu.trading.model.Coin;
-import com.basu.trading.model.Order;
-import com.basu.trading.model.OrderItem;
-import com.basu.trading.model.User;
+import com.basu.trading.model.*;
 import com.basu.trading.repository.OrderItemRepository;
 import com.basu.trading.repository.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,6 +24,9 @@ public class OrderServiceImpl implements OrderService{
 
     @Autowired
     private OrderItemRepository orderItemRepository;
+
+    @Autowired
+    private AssetService assetService;
 
     @Override
     public Order createOrder(User user, OrderItem orderItem, OrderType orderType) {
@@ -87,7 +87,16 @@ public class OrderServiceImpl implements OrderService{
         Order savedOrder=orderRepository.save(order);
 
         //create asset
+        Asset oldAsset = assetService.findAssetByUserIdAndCoinId(
+                order.getUser().getId(),
+                order.getOrderItem().getCoin().getId());
 
+        if (oldAsset==null){
+            assetService.createAsset(user, orderItem.getCoin(), orderItem.getQuantity());
+
+        }else {
+            assetService.updateAsset(oldAsset.getId(),quantity);
+        }
         return savedOrder;
     }
 
@@ -98,29 +107,41 @@ public class OrderServiceImpl implements OrderService{
         }
         double sellPrice = coin.getCurrentPrice();
 
-        double buyPrice = assetToSell.getPrice();
+        Asset assetToSell = assetService.findAssetByUserIdAndCoinId(user.getId(),coin.getId());
 
-        OrderItem orderItem = createOrderItem(coin,quantity,buyPrice,sellPrice);
 
-        Order order = createOrder(user,orderItem, OrderType.SELL);
-        orderItem.setOrder(order);
+        double buyPrice = assetToSell.getBuyPrice();
 
-        if (assetToSell.getQuantity()>=quantity){
+        if (assetToSell != null) {
 
-            order.setStatus(OrderStatus.SUCCESS);
-            order.setOrderType(OrderType.SELL);
-            Order savedOrder=orderRepository.save(order);
+            OrderItem orderItem = createOrderItem(coin, quantity, buyPrice, sellPrice);
 
-            walletService.payOrderPayment(order,user);
 
-            Asset updatedAsset = assetService.updatedAsset(assetToSell.getId(),-quantity);
-            if (updatedAsset.getQuantity()*coin.getCurrentPrice()<=1){
-                assetService.deleteAsset(updatedAsset.getId());
+            Order order = createOrder(user, orderItem, OrderType.SELL);
+            orderItem.setOrder(order);
+
+            if (assetToSell.getQuantity() >= quantity) {
+
+                order.setStatus(OrderStatus.SUCCESS);
+                order.setOrderType(OrderType.SELL);
+                Order savedOrder = orderRepository.save(order);
+
+                walletService.payOrderPayment(order, user);
+
+                Asset updatedAsset = assetService.updateAsset(
+                        assetToSell.getId(),
+                        -quantity
+                );
+
+                if (updatedAsset.getQuantity() * coin.getCurrentPrice() <= 1) {
+                    assetService.deleteAsset(updatedAsset.getId());
+                }
+                return savedOrder;
             }
-            return savedOrder;
-        }
 
-        throw new Exception("Insufficient quantity to sell");
+            throw new Exception("Insufficient quantity to sell");
+        }
+        throw new Exception("Asset not found");
     }
 
     @Override
